+++ wave_gtsam/src/preint_imu_factor.cpp
@@ -1,4 +1,4 @@
-#include "wave/gtsam/preintg_imu_factor/preint_imu_factor.hpp"
+#include "wave/gtsam/preint_imu_factor.hpp"
 
 namespace wave {
